Prototype of glance.
This commit provides a rough prototype of glance's functionality, based on
@thesarahcassidy's mockup (see `README`). There is a toolbar button (a "page
action") which is enabled on all LinkedIn pages.  When clicked, it shows a
popup with a very rough guess at the gender breakdown of the people listed
on the page.

Gender identity is a complex and multifaceted thing -- certainly far more
complex than checking a name against a list. The goal of this code is not to
make accurate assertions about any individual's gender, but instead to
provide a rough guess at the aggregate gender breakdown of the people
working at a given company.

In addition, this prototype makes the very English-language-centric
assumption that names can be broken down on spaces, and the first part is
the person's given name.

The prototype has three main components:

- `linkedin.js` is injected into all pages on LinkedIn. On page load, it
  sends a message to the extension instructing it to enable the page action.
  It also listens for messages instructing it to extract a list of names
  from the page.

- `popup.html` / `popup.js` are the user interface that pops up when the
  browser toolbar button is clicked. On page load, it sends a message to the
  extension instructing it to guess at the gender breakdown of the names on
  the page.

- `background.js` orchestrates messages between the other components, and
  also does the guess work.